// util/constants.d.ts
export declare const CLIPBOARD_FORMAT = "text/json";
export declare const DELIMITERS: string[];
export declare const COMPACT_HEIGHT = 24;
export declare const NORMAL_HEIGHT = 30;
export declare const LARGE_HEIGHT = 36;
export declare const COMPACT_PILL_HEIGHT = 18;
export declare const NORMAL_PILL_HEIGHT = 22;
export declare const LARGE_PILL_HEIGHT = 26;
export declare const KeyBoardkeys: {
    ArrowRight: string;
    ArrowLeft: string;
    ArrowUp: string;
    ArrowDown: string;
    PageUp: string;
    PageDown: string;
    Home: string;
    End: string;
    Enter: string;
    Tab: string;
    c: string;
    C: string;
    x: string;
    X: string;
    v: string;
    V: string;
    z: string;
    Z: string;
    Space: string;
    Escape: string;
};